auth_service/requirements.txt
fastapi[all]
pydantic
pydantic[email]
uvicorn
jinja2
pydantic_settings
aiosqlite
asyncpg
alembic
SQLAlchemy
bcrypt>=4.0.0
python-jose[cryptography]
loguru
aiofiles
redis
